Add support of Download Agent Authentication
When Download Agent Authentication (DAA) is enabled on the device,
several extra steps are needed in order to boot on the lk (DA).

We need to load and send:
- lk signature (default: lk.bin.sign)
- authentication file (default: auth_sv5.auth)

The user can specify the file names with commandline arguments:
$ bootrom-tool -s lk.bin.sign -t auth_sv5.auth

These files are not required if DAA is disabled on the device.

The version has also been bumped to 1.1.3

Signed-off-by: Julien Masson <jmasson@baylibre.com>